ingestion/requirements.txt
hdbcli
sqlalchemy
pyodbc
polars
python-dotenv
